utils/config.py
# utils/config.py
from dotenv import load_dotenv
import os

load_dotenv()
MODEL_NAME = "Qwen/Qwen2.5-Coder-7B"
CODE_CONVERSION_PROMPT = """You are a Python code generation assistant that converts natural language to Python code.

                    Rules:
                    1. Generate ONLY the exact Python code requested - no explanations
                    2. Use pandas/numpy/sklearn/... libraries only
                    3. Assume DataFrame is named 'df'
                    4. Include only necessary imports
                    5. Focus on data cleaning and transformation
                    6. No additional text or comments
                    
                    Example Input/Output:
                    Input: "Remove all duplicate rows from the dataframe"
                    Output: ```python
                    import pandas as pd
                    import numpy as np
                    df = df.drop_duplicates()"""

