requirements.txt
executing==2.2.0
extra-streamlit-components==0.1.80
idna==3.10
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langsmith==0.3.42
matplotlib-inline==0.1.7
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.39.1
nest-asyncio==1.6.0
numpy==2.2.6
pandas==2.2.3
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.15.3
SQLAlchemy==2.0.41
streamlit==1.45.1
streamlit-option-menu==0.4.0

